requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Django==3.2.16
django-redis==5.2.0
ecdsa==0.18.0
et-xmlfile==1.1.0
openpyxl==3.1.0
Pillow==9.4.0
pyasn1==0.4.8
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2022.7.1
redis==4.5.4
rsa==4.9
six==1.16.0
sqlparse==0.4.3
xlrd==2.0.1
xpinyin==0.7.6
django-bulk-update-or-create==0.3.0